+++ action_server/src/sema4ai/action_server/_actions_import.py
@@ -62,7 +62,7 @@ def _log_deprecated_conda():
             "Deprecated: The file for defining the environment is now `package.yaml`.\n"
             "It's not a one to one mapping for action-server.yaml, but\n"
             f"`{cmd} package update` can be used to make most of the needed changes.\n"
-            "See: https://github.com/robocorp/robocorp/blob/master/action_server/docs/guides/01-package-yaml.md for more details."
+            "See: https://github.com/Sema4AI/actions/blob/master/action_server/docs/guides/01-package-yaml.md for more details."
         )
     )