+++ action_server/src/sema4ai/action_server/_rcc.py
@@ -39,9 +39,9 @@ def as_str(s) -> str:
 
 
 class Rcc(object):
-    def __init__(self, rcc_location: Path, robocorp_home: Optional[Path]):
+    def __init__(self, rcc_location: Path, sema4ai_home: Optional[Path]):
         self._rcc_location = rcc_location
-        self._robocorp_home = robocorp_home
+        self.sema4ai_home = sema4ai_home
         self.config_location = os.environ.get(
             "S4_ACTION_SERVER_RCC_CONFIG_LOCATION", ""
         )
@@ -54,8 +54,9 @@ class Rcc(object):
         env["PYTHONIOENCODING"] = "utf-8"
         env["PYTHONUNBUFFERED"] = "1"
 
-        if self._robocorp_home:
-            env["ROBOCORP_HOME"] = str(self._robocorp_home)
+        if self.sema4ai_home:
+            env["SEMA4AI_HOME"] = str(self.sema4ai_home)
+
         return env
 
     def _compute_launch_args_and_kwargs(
@@ -68,7 +69,12 @@ class Rcc(object):
 
         kwargs: dict = build_subprocess_kwargs(cwd, env, stderr=stderr)
         rcc_location = str(self._rcc_location)
-        args = [rcc_location] + args + ["--controller", "action-server", "--bundled"]
+        args = (
+            [rcc_location]
+            + args
+            + ["--controller", "action-server", "--bundled", "--sema4ai"]
+        )
+
         return args, kwargs
 
     def _run_rcc(
@@ -87,7 +93,7 @@ class Rcc(object):
         Returns an ActionResult where the result is the stdout of the executed command.
 
         :param log_errors:
-            If false, errors won't be logged (i.e.: should be false when errors
+            If false, errors won't be logged (i.e.: should be false when errors-+
             are expected).
 
         :param stderr:
@@ -99,9 +105,9 @@ class Rcc(object):
         from sema4ai.action_server._robo_utils.process import check_output_interactive
 
         env = self._compute_env()
-        robocorp_home = env.get("ROBOCORP_HOME")
-        if not robocorp_home:
-            robocorp_home = "<unset>"
+        sema4ai_home = env.get("SEMA4AI_HOME")
+        if not sema4ai_home:
+            sema4ai_home = "<unset>"
 
         args, kwargs = self._compute_launch_args_and_kwargs(cwd, env, args, stderr)
         cmdline = list2cmdline([str(x) for x in args])
@@ -146,8 +152,9 @@ class Rcc(object):
         except CalledProcessError as e:
             stdout = as_str(e.stdout)
             stderr = as_str(e.stderr)
+
             msg = (
-                f"Error running: {cmdline}.\nROBOCORP_HOME: {robocorp_home}\n\n"
+                f"Error running: {cmdline}.\nSEMA4AI_HOME: {sema4ai_home}\n\n"
                 f"Stdout: {stdout}\nStderr: {stderr}"
             )
             if hide_in_log:
@@ -362,14 +369,14 @@ _rcc: Optional["Rcc"] = None
 
 
 @contextmanager
-def initialize_rcc(rcc_location: Path, robocorp_home: Optional[Path]) -> Iterator[Rcc]:
+def initialize_rcc(rcc_location: Path, sema4ai_home: Optional[Path]) -> Iterator[Rcc]:
     global _rcc
 
     if _rcc:
         yield _rcc
         return
 
-    rcc = Rcc(rcc_location, robocorp_home)
+    rcc = Rcc(rcc_location, sema4ai_home)
     _rcc = rcc
     try:
         yield rcc